third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-get-basics.html

<!doctype html>
<title>Credential Manager: get() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/credentialmanager/credential_manager.mojom.js"></script>
<script src="/gen/third_party/WebKit/public/platform/modules/webauth/authenticator.mojom.js"></script>
<script src="resources/credential-helpers.js"></script>
<script>

add_completion_callback(() => {
  mockCredentialManager.reset();
});

promise_test(_ => {
  return navigator.credentials.get().then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get() with no argument.");

promise_test(_ => {
  return navigator.credentials.get({}).then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get({}).");

promise_test(_ => {
  return navigator.credentials.get({
    federated: {
      providers: [ 'https://example.com/' ]
    }
  }).then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get() with a valid options including FederatedCredentialRequestOptions.");

promise_test(_ => {
  return navigator.credentials.get({
    password: true,
    unmediated: true
  }).then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get() with a valid options including password and unmediated.");

promise_test(_ => {
  return navigator.credentials.get({
    federated: {
      providers: [ 'https://example.com/' ]
    },
    unmediated: true
  }).then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get() with a valid options including federated and unmediated.");

promise_test(_ => {
  return navigator.credentials.get({
    password: true,
    federated: {
      providers: [ 'https://example.com/' ]
    },
    unmediated: true
  }).then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get() with a valid options including federated, password and unmediated.");

promise_test(_ => {
  return navigator.credentials.get({
    unmediated: true
  }).then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get() with a valid options including unmediated.");

promise_test(_ => {
  return navigator.credentials.get({
    mediation: "silent"
  }).then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get() with a valid options including mediation.");

promise_test(_ => {
  return navigator.credentials.get({
    notValid: 'yay!'
  }).then(r => {
    assert_equals(r, null);
  });
}, "navigator.credentials.get() with an options including an unknown attribute.");

promise_test(_ => {
  var id = "id";
  var password = "pencil";
  var name = "name";
  var icon = "http://example.com/";

  mockCredentialManager.setResponse(id, password, name, icon);

  return navigator.credentials.get({
    password: true
  }).then(r => {
    assert_equals(r.id, id, "id");
    assert_equals(r.password, password, "password");
    assert_equals(r.name, name, "name");
    assert_equals(r.iconURL, icon, "icon");
  });
}, "Verify that the mock returns the values we give it.");

</script>
